==> Kaggle/Titanic/01_Titanic Random Forest.ipynb <==
# Import Moudles
---

Import CPU vs GPU all code

If you want to see how many utilize your gpu, run below on terminal

```cmd
watch -n 0.1 nvidia-smi
```

import time

import cupy as cp
import numpy as np

import cudf
import pandas as pd

from sklearn.ensemble import RandomForestClassifier as RF_cpu
from cuml.ensemble    import RandomForestClassifier as RF_gpu

# Make Timer Class
---

Always Check Time 

CPU based code vs GPU based code

class Timer:
    def __init__(self, name="cpu"):
        self.name = name
        
    def __enter__(self):
        self.start = time.perf_counter()
        
    def __exit__(self, type, value, trackback):
        self.end = time.perf_counter()
        self.execute_time = self.end - self.start
        print(f"{self.name} execute time : {self.execute_time:.4f} seconds")

# Load Data

import os
import glob

path = glob.glob(os.path.join("data", "*"))
print(path)

with Timer(name="cpu") as cpu_time:
    train_data_cpu = pd.read_csv("./data/train.csv")
    test_data_cpu  = pd.read_csv("./data/test.csv")

with Timer(name="gpu") as gpu_time:
    train_data_gpu = cudf.read_csv("./data/train.csv")
    test_data_gpu  = cudf.read_csv("./data/test.csv")

train_data_cpu.head()

train_data_gpu.head()

with Timer(name="cpu") as cpu_time:
    women = train_data_cpu.loc[train_data_cpu.Sex == 'female']["Survived"]
    rate_women = sum(women)/len(women)    
    print(f"% of women who survived: {rate_women}")
    
    men = train_data_cpu.loc[train_data_cpu.Sex == 'male']["Survived"]
    rate_men = sum(men)/len(men)    
    print(f"% of men who survived: {rate_men}")

with Timer(name="gpu") as gpu_time:
    women = train_data_gpu.loc[train_data_gpu.Sex == 'female']["Survived"]
    rate_women = women.sum()/len(women)    
    print(f"% of women who survived: {rate_women}")
    
    men = train_data_gpu.loc[train_data_gpu.Sex == 'male']["Survived"]
    rate_men = men.sum()/len(men)    
    print(f"% of men who survived: {rate_men}")

with Timer(name="cpu run") as cpu_time:
    y = train_data_cpu["Survived"]
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = pd.get_dummies(train_data_cpu[features]).astype(np.float32)
    X_test = pd.get_dummies(test_data_cpu[features]).astype(np.float32)
    
    model = RF_cpu(n_estimators=100, max_depth=5, random_state=1)
    model.fit(X, y)
    predictions = model.predict(X_test)
    
    output = pd.DataFrame({'PassengerId': test_data_cpu.PassengerId, 'Survived': predictions})
    
with Timer(name="cpu save") as cpu_time:
    output.to_csv('submission_RF_cpu.csv', index=False)
    print("Your submission was successfully saved!")

with Timer(name="gpu run") as gpu_time:
    y = train_data_gpu["Survived"]
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = cudf.get_dummies(train_data_gpu[features]).astype(cp.float32)
    X_test = cudf.get_dummies(test_data_gpu[features]).astype(cp.float32)
    
    model = RF_gpu(n_estimators=100, max_depth=5, random_state=1, n_streams=1)
    model.fit(X, y)
    predictions = model.predict(X_test)
    
    output = cudf.DataFrame({'PassengerId': test_data_gpu.PassengerId, 'Survived': predictions})
    
with Timer(name="gpu save") as gpu_time:    
    output.to_csv('submission_RF_gpu.csv', index=False)
    print("Your submission was successfully saved!")

# More complicate and iterate samples

with Timer(name="cpu run") as cpu_time:
    y = train_data_cpu["Survived"]
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = pd.get_dummies(train_data_cpu[features]).astype(np.float32)
    X_test = pd.get_dummies(test_data_cpu[features]).astype(np.float32)
    
    model = RF_cpu(n_estimators=10000, max_depth=100, random_state=1)
    model.fit(X, y)
    predictions = model.predict(X_test)
    
    output = pd.DataFrame({'PassengerId': test_data_cpu.PassengerId, 'Survived': predictions})

with Timer(name="cpu save") as cpu_time:
    output.to_csv('submission_RF_cpu.csv', index=False)
    print("Your submission was successfully saved!")

with Timer(name="gpu run") as cpu_time:
    y = train_data_gpu["Survived"]
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = cudf.get_dummies(train_data_gpu[features]).astype(cp.float32)
    X_test = cudf.get_dummies(test_data_gpu[features]).astype(cp.float32)
    
    model = RF_gpu(n_estimators=10000, max_depth=100, random_state=1, n_streams=1)
    model.fit(X, y)
    predictions = model.predict(X_test)
    
    output = cudf.DataFrame({'PassengerId': test_data_gpu.PassengerId, 'Survived': predictions})

with Timer(name="gpu save") as cpu_time:
    output.to_csv('submission_RF_gpu.csv', index=False)
    print("Your submission was successfully saved!")
