==> Kaggle/Titanic/02_Titanic Random Forest with Dask.ipynb <==
# Import Moudles
---

Import CPU vs GPU all code

If you want to see how many utilize your gpu, run below on terminal

```cmd
watch -n 0.1 nvidia-smi
```

Dask Visualizer need to install graphvis jupyter-cytoscape

```cmd
conda install graphvis ipycytoscape -y
```

import time

import cupy as cp

import cudf
import dask
import dask_cudf
import dask.dataframe as dd
dask.config.set({"dataframe.backend": "cudf"})

from cuml.ensemble    import RandomForestClassifier as RF

class Timer:
    def __init__(self, name="cpu"):
        self.name = name
        
    def __enter__(self):
        self.start = time.perf_counter()
        
    def __exit__(self, type, value, trackback):
        self.end = time.perf_counter()
        self.execute_time = self.end - self.start
        print(f"{self.name} execute time : {self.execute_time:.4f} seconds")

import os
import glob

path = glob.glob(os.path.join("data", "*"))
print(path)

with Timer(name="gpu read csv") as gpu_time:
    train_data_gpu = cudf.read_csv("./data/train.csv")
    test_data_gpu  = cudf.read_csv("./data/test.csv")

with Timer(name="gpu to parquet") as gpu_time:
    train_data_gpu.to_parquet("./data/train.parquet")
    test_data_gpu.to_parquet("./data/test.parquet")



with Timer(name="dask read parquet") as dask_time:
    train_dask = dd.read_parquet("./data/train.parquet")
    test_dask  = dd.read_parquet("./data/test.parquet")

train_dask.visualize()

test_dask.visualize()

# Compute vs Persist

with Timer(name="gpu") as gpu_time:
    women = train_dask.loc[train_dask['Sex'] == 'female']["Survived"]
    rate_women = women.sum()/len(women)    
    print(f"% of women who survived: {rate_women.compute()}")
    
    men = train_dask.loc[train_dask['Sex'] == 'male']["Survived"]
    rate_men = men.sum()/len(men)    
    print(f"% of men who survived: {rate_men.compute()}")

rate_men.visualize()

train_dask = train_dask.persist()
test_dask  = test_dask.persist()

with Timer(name="gpu") as gpu_time:
    women = train_dask.loc[train_dask['Sex'] == 'female']["Survived"]
    rate_women = women.sum()/len(women)    
    print(f"% of women who survived: {rate_women.compute()}")
    
    men = train_dask.loc[train_dask['Sex'] == 'male']["Survived"]
    rate_men = men.sum()/len(men)    
    print(f"% of men who survived: {rate_men.compute()}")

rate_men.visualize()

with Timer(name="gpu run") as gpu_time:
    y = train_dask["Survived"].persist()
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = dd.reshape.get_dummies(train_dask[features].categorize()).astype(cp.float32)
    X_test = dd.reshape.get_dummies(test_dask[features].categorize()).astype(cp.float32)
    
    model = RF(n_estimators=100, max_depth=5, random_state=1, n_streams=1)
    model.fit(X, y)
    predictions = model.predict(X_test.compute())
    
    output = cudf.DataFrame({'PassengerId': test_data_gpu.PassengerId, 'Survived': predictions})
    
with Timer(name="gpu save") as gpu_time:    
    output.to_csv('submission_RF_dask.csv', index=False)
    print("Your submission was successfully saved!")

with Timer(name="gpu run") as gpu_time:
    y = train_dask["Survived"].persist()
    
    features = ["Pclass", "Sex", "SibSp", "Parch"]
    X = dd.reshape.get_dummies(train_dask[features].categorize()).astype(cp.float32)
    X_test = dd.reshape.get_dummies(test_dask[features].categorize()).astype(cp.float32)
    
    model = RF(n_estimators=10000, max_depth=100, random_state=1, n_streams=1)
    model.fit(X, y)
    predictions = model.predict(X_test.compute())
    
    output = cudf.DataFrame({'PassengerId': test_data_gpu.PassengerId, 'Survived': predictions})
    
with Timer(name="gpu save") as gpu_time:    
    output.to_csv('submission_RF_dask.csv', index=False)
    print("Your submission was successfully saved!")
